+++ resources/views/requirements/index.blade.php
@@ -14,8 +14,9 @@
     <tr>
         <th width="5%">#</th>
         <th width="20%">Title</th>
-        <th width="45%">Description</th>
-        <th width="15%">Created On</th>
+        <th width="40%">Description</th>
+        <th width="10%">Created By</th>
+        <th width="10%">Created On</th>
         <th width="15%">Action</th>
     </tr>
     </thead>
@@ -26,16 +27,21 @@
             <th scope="row">{{ $i }}</th>
             <td><a href="{{ url('/requirements', [$req->id]) }}">{{ $req->title }}</a></td>
             <td>{{ $req->description }}</td>
+            <td>{{ $req->created_by }}</td>
             <td>{{ $req->created_at }}</td>
             <td>
-                <a href="{{ url('/requirements', [$req->id, 'edit']) }}" title="Edit!" class="btn btn-info btn-sm" style="float: left; margin-right: 5px;">
-                    <span class="glyphicon glyphicon-edit" aria-hidden="true"></span> Edit
-                </a>
-                {!! Form::open(['method' => 'DELETE', 'route' => ['requirements.destroy', $req->id], 'onsubmit' => 'return confirm("Are you sure you want to delete?")']) !!}
-                <button type="submit" class="btn btn-danger btn-sm">
-                    <span class="glyphicon glyphicon-remove" aria-hidden="true"></span> Delete
-                </button>
-                {!! Form::close() !!}
+                @if($req->user_id === $authUser->id)
+                    <a href="{{ url('/requirements', [$req->id, 'edit']) }}" title="Edit!" class="btn btn-info btn-sm" style="float: left; margin-right: 5px;">
+                        <span class="glyphicon glyphicon-edit" aria-hidden="true"></span> Edit
+                    </a>
+                    {!! Form::open(['method' => 'DELETE', 'route' => ['requirements.destroy', $req->id], 'onsubmit' => 'return confirm("Are you sure you want to delete?")']) !!}
+                    <button type="submit" class="btn btn-danger btn-sm">
+                        <span class="glyphicon glyphicon-remove" aria-hidden="true"></span> Delete
+                    </button>
+                    {!! Form::close() !!}
+                @else
+                    <span>--</span>
+                @endif
             </td>
         </tr>
         <?php $i++; ?>
